Win32 build - copy of code

+++ src/output.c
@@ -100,9 +100,6 @@ YYNTBASE = ntokens.
 
 #include <limits.h>
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <malloc.h>
 #include "system.h"
 #include "machine.h"
 #include "new.h"
@@ -502,7 +499,7 @@ void
 output_rule_data()
 {
   register int i;
-  register size_t j;
+  register int j;
 
   fprintf(ftable,
      "\n#if YY_%s_DEBUG != 0\nstatic const short yyrline[] = { 0",parser_name);
@@ -1418,7 +1415,7 @@ FILE *fin,*fout;
 	  {if (c == '$')
 	    {
   	    if (!nolinesflag)
-   		{fprintf(fout,
+		{fprintf(fout,
                 "\n/* #line %d \"%s\" */\n#line @\n",
 		 (*pcounter), quoted_filename(fil_name));
                 }
